requirements.txt
Babel==2.5.1
Beaker==1.9.0
Django==1.9.5
Pillow==4.2.1
Pyoko==0.9.2
SpiffWorkflow==v0.4
Unidecode==0.04.21
Werkzeug==0.12.2
argparse==1.2.1
backports-abc==0.5
beaker-extensions==0.1.1dev
billiard==3.5.0.3
celery==4.1.0
certifi==2017.07.27.1
django-extensions==1.9.1
django-filer==1.2.8
django-money==0.11.4
django-mptt==0.8.7
django-polymorphic==1.1b2
easy-thumbnails==2.3
enum34==1.1.6
falcon==1.3.0
kombu==4.1.0
lazy-object-proxy==1.3.1
mercurial==3.1
numpy==1.11.0
olefile==0.44
passlib==1.7.1
pika==0.11.0
py-moneyed==0.7.0
python-mimeparse==1.6.0
pytz==2016.3
redis==2.10.6
riak==2.7.0
singledispatch==3.4.0.3
six==1.11.0
tornado==4.5.2
wsgiref==0.1.2
zengine==0.7.9
